Test cases initialized for CrudColumn
